consumer/requirements.txt
kafka-python-ng>=2.2.3
boto3==1.34.0